chore(core): typehinting more lib functions
ref #14075

+++ mod/blog/views/default/resources/blog/add.php
@@ -2,7 +2,7 @@
 
 use Elgg\Exceptions\Http\EntityPermissionsException;
 
-$guid = elgg_extract('guid', $vars);
+$guid = (int) elgg_extract('guid', $vars);
 if (!$guid) {
 	$guid = elgg_get_logged_in_user_guid();
 }
