+++ mod/bookmarks/views/default/resources/bookmarks/group.php
@@ -1,6 +1,6 @@
 <?php
 
-$group_guid = elgg_extract('guid', $vars);
+$group_guid = (int) elgg_extract('guid', $vars);
 
 elgg_entity_gatekeeper($group_guid, 'group');
