+++ mod/groups/views/default/groups/invite/user_html.php
@@ -22,7 +22,7 @@ $delete_icon = elgg_view_icon('delete', ['class' => 'elgg-autocomplete-item-remo
 
 $title = $entity->getDisplayName();
 
-$group = get_entity(elgg_extract('group_guid', $vars));
+$group = get_entity((int) elgg_extract('group_guid', $vars));
 if ($group instanceof ElggGroup) {
 	if ($group->isMember($entity)) {
 		$title .= elgg_format_element('span', ['class' => ['mls', 'elgg-subtext']], elgg_echo('groups:invite:member'));
