+++ mod/thewire/views/default/resources/thewire/view.php
@@ -3,7 +3,7 @@
  * View individual wire post
  */
 
-$guid = elgg_extract('guid', $vars);
+$guid = (int) elgg_extract('guid', $vars);
 elgg_entity_gatekeeper($guid, 'object', 'thewire');
 
 /* @var $post ElggWire */
